Implemented review comments, splitting into two rules, config call clean up, etc.
