+++ config/config.default.yaml
@@ -252,9 +252,7 @@ transmission:
       efficiency_static: 0.97
   hydrogen:
     enable: true
-    gabriel_filter:
-      enable: true
-      min_degree: 1
+    gabriel_filter_min_degree: 1
     max_offshore_haversine_distance: .inf
     length_factor: 1.25
     cost_factor: 1
@@ -267,9 +265,7 @@ transmission:
       compression_per_1000km: 0.018
   carbon_dioxide:
     enable: true
-    gabriel_filter:
-      enable: true
-      min_degree: 1
+    gabriel_filter_min_degree: 1
     max_offshore_haversine_distance: .inf
     length_factor: 1.25
     cost_factor: 1
