+++ rules/build_sector.smk
@@ -148,59 +148,46 @@ rule cluster_gas_network:
         scripts("cluster_gas_network.py")
 
 
-def transmission_candidate_specs() -> list[tuple[int, str]]:
-    specs = set()
-
-    for carrier_config in config.get("transmission", {}).values():
-        if not isinstance(carrier_config, dict):
-            continue
-        if not carrier_config.get("enable", False):
-            continue
-        gabriel_filter = carrier_config.get("gabriel_filter", {})
-        if not isinstance(gabriel_filter, dict):
-            continue
-        if not gabriel_filter.get("enable", False):
-            continue
-        if "min_degree" not in gabriel_filter:
-            continue
-
-        min_degree = int(gabriel_filter["min_degree"])
-        max_offdist = f"{float(carrier_config.get('max_offshore_haversine_distance', float('inf'))):g}"
-        specs.add((min_degree, max_offdist))
-
-    return sorted(specs, key=lambda spec: (spec[0], float(spec[1])))
-
-
-rule build_transmission_topology:
+rule build_transmission_delaunay_graph:
     input:
         network=resources("networks/base_s_{clusters}.nc"),
         offshore_shapes=resources("offshore_shapes.geojson"),
     output:
-        all_edges=resources("transmission/all_edges_{clusters}.geojson"),
-        candidates=[
-            resources(
-                "transmission/candidates_{clusters}_min_"
-                + f"{int(min_degree)}_maxoffdist_{max_offdist}_km.geojson"
-            )
-            for min_degree, max_offdist in transmission_candidate_specs()
-        ],
+        delaunay_graph=resources("transmission/delaunay_graph_{clusters}.geojson"),
     log:
-        logs("build_transmission_topology_{clusters}.log"),
+        logs("build_transmission_delaunay_graph_{clusters}.log"),
     benchmark:
-        benchmarks("build_transmission_topology/{clusters}")
+        benchmarks("build_transmission_delaunay_graph/{clusters}")
     resources:
         mem_mb=4000,
-    params:
-        transmission=config_provider("transmission"),
-        candidate_specs=[
-            {
-                "min_degree": int(min_degree),
-                "max_offshore_haversine_distance_km": float(max_offdist),
-            }
-            for min_degree, max_offdist in transmission_candidate_specs()
-        ],
     message:
-        "Building transmission candidates for {wildcards.clusters} clusters"
+        "Building Delaunay triangulation graph for {wildcards.clusters} clusters"
+    script:
+        scripts("build_transmission_delaunay_graph.py")
+
+
+rule build_transmission_topology:
+    input:
+        delaunay_graph=resources("transmission/delaunay_graph_{clusters}.geojson"),
+    output:
+        candidates=resources(
+            "transmission/candidates_{clusters}_min_{min_degree}_maxoffdist_{max_offdist}_km.geojson"
+        ),
+    log:
+        logs(
+            "build_transmission_topology_{clusters}_min_{min_degree}_maxoffdist_{max_offdist}.log"
+        ),
+    benchmark:
+        benchmarks(
+            "build_transmission_topology/{clusters}_min_{min_degree}_maxoffdist_{max_offdist}"
+        )
+    resources:
+        mem_mb=2000,
+    params:
+        min_degree=lambda w: int(w.min_degree),
+        max_offdist=lambda w: float(w.max_offdist),
+    message:
+        "Building transmission topology for {wildcards.clusters} clusters (min_degree={wildcards.min_degree}, max_offdist={wildcards.max_offdist} km)"
     script:
         scripts("build_transmission_topology.py")
 
@@ -1624,21 +1611,21 @@ def input_heat_source_power(w):
 def input_transmission_candidates(w):
     """Generate transmission candidate file inputs for enabled transmission carriers."""
     candidates = {}
-    for carrier, carrier_cfg in config.get("transmission", {}).items():
-        if not isinstance(carrier_cfg, dict) or not carrier_cfg.get("enable", False):
-            continue
-        gabriel_cfg = carrier_cfg.get("gabriel_filter")
-        if not isinstance(gabriel_cfg, dict):
+    for carrier in ("hydrogen", "carbon_dioxide"):
+        carrier_cfg = config["transmission"][carrier]
+        if not carrier_cfg["enable"]:
             continue
         key = f"{carrier}_transmission_candidates"
-        if gabriel_cfg.get("enable", False) and "min_degree" in gabriel_cfg:
-            min_degree = int(gabriel_cfg["min_degree"])
-            max_offdist = f"{float(carrier_cfg.get('max_offshore_haversine_distance', float('inf'))):g}"
+        min_degree = carrier_cfg["gabriel_filter_min_degree"]
+        max_offdist = carrier_cfg["max_offshore_haversine_distance"]
+        if min_degree > 0 or max_offdist < float("inf"):
             candidates[key] = resources(
                 f"transmission/candidates_{{clusters}}_min_{min_degree}_maxoffdist_{max_offdist}_km.geojson"
             )
         else:
-            candidates[key] = resources("transmission/all_edges_{clusters}.geojson")
+            candidates[key] = resources(
+                "transmission/delaunay_graph_{clusters}.geojson"
+            )
     return candidates
